+++ lenstronomy/GalKin/galkin_multiobservation.py
@@ -54,14 +54,14 @@ class GalkinMultiObservation(GalkinModel):
 
         for i in range(0, num_kin_sampling):
             r, R, x, y = self.numerics.draw_light(kwargs_light)
-            sigma2_R = self.numerics.sigma_s2(r, R, kwargs_mass, kwargs_light, kwargs_anisotropy)
+            sigma2_IR, IR = self.numerics.sigma_s2(r, R, kwargs_mass, kwargs_light, kwargs_anisotropy)
             for obs_index, observation in enumerate(self._observation_list):
                 for k in range(0, num_psf_sampling):
                     x_, y_ = observation.displace_psf(x, y)
                     bool, _ = observation.aperture_select(x_, y_)
                     if bool is True:
-                        sigma2_R_sum[obs_index] += sigma2_R
-                        count_draws[obs_index] += 1
+                        sigma2_R_sum[obs_index] += sigma2_IR
+                        count_draws[obs_index] += IR
 
         sigma_s2_average = sigma2_R_sum / count_draws
         # apply unit conversion from arc seconds and deflections to physical velocity dispersion in (km/s)