+++ test/test_GalKin/test_galkin.py
@@ -3,115 +3,20 @@ Tests for `galkin` module.
 """
 import pytest
 import unittest
+import copy
 import numpy.testing as npt
 import numpy as np
 import scipy.integrate as integrate
 from lenstronomy.GalKin.galkin import Galkin
 from lenstronomy.GalKin.light_profile import LightProfile
-from lenstronomy.GalKin.analytic_kinematics import AnalyticKinematics
 import lenstronomy.Util.param_util as param_util
+from lenstronomy.Util import constants as const
 
 
 class TestGalkin(object):
 
     def setup(self):
         np.random.seed(42)
-
-    def test_log_linear_integral(self):
-        # light profile
-        light_profile_list = ['HERNQUIST']
-        Rs = .5
-        kwargs_light = [{'Rs':  Rs, 'amp': 1.}]  # effective half light radius (2d projected) in arcsec
-        # 0.551 *
-        # mass profile
-        mass_profile_list = ['SPP']
-        theta_E = 1.2
-        gamma = 2.
-        kwargs_profile = [{'theta_E': theta_E, 'gamma': gamma}]  # Einstein radius (arcsec) and power-law slope
-
-        # anisotropy profile
-        anisotropy_type = 'OM'
-        r_ani = 2.
-        kwargs_anisotropy = {'r_ani': r_ani}  # anisotropy radius [arcsec]
-
-        # aperture as slit
-        aperture_type = 'slit'
-
-        psf_fwhm = 0.7  # Gaussian FWHM psf
-        kwargs_cosmo = {'d_d': 1000, 'd_s': 1500, 'd_ds': 800}
-        kwargs_numerics_linear = {'interpol_grid_num': 500, 'log_integration': False,
-                           'max_integrate': 10, 'min_integrate': 0.001}
-        kwargs_numerics_log = {'interpol_grid_num': 500, 'log_integration': True,
-                           'max_integrate': 10, 'min_integrate': 0.001}
-        kwargs_aperture = {'width': 1, 'length': 1., 'aperture_type': aperture_type}
-        kwargs_psf = {'psf_type': 'GAUSSIAN', 'fwhm': psf_fwhm}
-        kwargs_model = {'mass_profile_list': mass_profile_list,
-                        'light_profile_list': light_profile_list,
-                        'anisotropy_model': anisotropy_type}
-        galkin_linear = Galkin(kwargs_model=kwargs_model, kwargs_aperture=kwargs_aperture,
-                               kwargs_psf=kwargs_psf, kwargs_cosmo=kwargs_cosmo, kwargs_numerics=kwargs_numerics_linear)
-        galkin_log = Galkin(kwargs_model=kwargs_model, kwargs_aperture=kwargs_aperture,
-                            kwargs_psf=kwargs_psf, kwargs_cosmo=kwargs_cosmo, kwargs_numerics=kwargs_numerics_log)
-        R = np.linspace(0.05, 1, 100)
-        lin_I_R = np.zeros_like(R)
-        log_I_R = np.zeros_like(R)
-        for i in range(len(R)):
-            lin_I_R[i] = galkin_linear.numerics._I_R_sigma2(R[i], kwargs_profile, kwargs_light, kwargs_anisotropy)
-            log_I_R[i] = galkin_log.numerics._I_R_sigma2(R[i], kwargs_profile, kwargs_light, kwargs_anisotropy)
-        print(log_I_R/lin_I_R)
-        for i in range(len(R)):
-            npt.assert_almost_equal(log_I_R[i] / lin_I_R[i], 1, decimal=2)
-
-    def test_log_vs_linear_integral(self):
-        """
-        here we test logarithmic vs linear integral in an end-to-end fashion.
-        We do not demand the highest level of precsions here!!!
-        We are using the luminosity-weighted velocitydispersion integration calculation in this test.
-        """
-
-        # light profile
-        light_profile_list = ['HERNQUIST']
-        Rs = .5
-        kwargs_light = [{'Rs':  Rs, 'amp': 1.}]  # effective half light radius (2d projected) in arcsec
-        # 0.551 *
-        # mass profile
-        mass_profile_list = ['SPP']
-        theta_E = 1.2
-        gamma = 2.
-        kwargs_profile = [{'theta_E': theta_E, 'gamma': gamma}]  # Einstein radius (arcsec) and power-law slope
-
-        # anisotropy profile
-        anisotropy_type = 'OM'
-        r_ani = 2.
-        kwargs_anisotropy = {'r_ani': r_ani}  # anisotropy radius [arcsec]
-
-        # aperture as slit
-        aperture_type = 'slit'
-        length = 3.8
-        width = 0.9
-        kwargs_aperture = {'aperture_type': aperture_type, 'length': length, 'width': width, 'center_ra': 0, 'center_dec': 0, 'angle': 0}
-
-        psf_fwhm = 0.7  # Gaussian FWHM psf
-        kwargs_cosmo = {'d_d': 1000, 'd_s': 1500, 'd_ds': 800}
-        kwargs_numerics_log = {'interpol_grid_num': 500, 'log_integration': True,
-                           'max_integrate': 10, 'lum_weight_int_method': True}
-        kwargs_numerics_linear = {'interpol_grid_num': 500, 'log_integration': False,
-                           'max_integrate': 10, 'lum_weight_int_method': True}
-        kwargs_psf = {'psf_type': 'GAUSSIAN', 'fwhm': psf_fwhm}
-        kwargs_model = {'mass_profile_list': mass_profile_list,
-                        'light_profile_list': light_profile_list,
-                        'anisotropy_model': anisotropy_type}
-        galkin_linear = Galkin(kwargs_model=kwargs_model, kwargs_aperture=kwargs_aperture, kwargs_psf=kwargs_psf,
-                               kwargs_cosmo=kwargs_cosmo, kwargs_numerics=kwargs_numerics_linear)
-
-        sigma_v = galkin_linear.dispersion(kwargs_profile, kwargs_light, kwargs_anisotropy, sampling_number=1000)
-        galkin_log = Galkin(kwargs_model=kwargs_model, kwargs_aperture=kwargs_aperture, kwargs_psf=kwargs_psf,
-                            kwargs_cosmo=kwargs_cosmo, kwargs_numerics=kwargs_numerics_log)
-        sigma_v2 = galkin_log.dispersion(kwargs_profile, kwargs_light, kwargs_anisotropy, sampling_number=1000)
-        print(sigma_v, sigma_v2, 'sigma_v linear, sigma_v log')
-        print((sigma_v/sigma_v2)**2)
-
-        npt.assert_almost_equal(sigma_v/sigma_v2, 1, decimal=1)
 
     def test_compare_power_law(self):
         """
@@ -142,28 +47,96 @@ class TestGalkin(object):
 
         psf_fwhm = 1.  # Gaussian FWHM psf
         kwargs_cosmo = {'d_d': 1000, 'd_s': 1500, 'd_ds': 800}
-        kwargs_numerics = {'interpol_grid_num': 500, 'log_integration': True,
-                           'max_integrate': 100}
+        kwargs_numerics = {'interpol_grid_num': 1000, 'max_integrate': 1000, 'min_integrate': 0.001}
         kwargs_model = {'mass_profile_list': mass_profile_list,
                         'light_profile_list': light_profile_list,
                         'anisotropy_model': anisotropy_type}
         kwargs_psf = {'psf_type': 'GAUSSIAN', 'fwhm': psf_fwhm}
-        galkin = Galkin(kwargs_model=kwargs_model, kwargs_aperture=kwargs_aperture, kwargs_psf=kwargs_psf,
-                        kwargs_cosmo=kwargs_cosmo, kwargs_numerics=kwargs_numerics)
-        sigma_v = galkin.dispersion(kwargs_profile, kwargs_light, kwargs_anisotropy, sampling_number=1000)
 
-        kwargs_numerics = {'interpol_grid_num': 500, 'log_integration': False, 'max_integrate': 10}
-
-        galkin = Galkin(kwargs_model=kwargs_model, kwargs_aperture=kwargs_aperture, kwargs_psf=kwargs_psf,
-                        kwargs_cosmo=kwargs_cosmo, kwargs_numerics=kwargs_numerics, analytic_kinematics=False)
-        sigma_v_lin = galkin.dispersion(kwargs_profile, kwargs_light, kwargs_anisotropy, sampling_number=1000)
-
-        los_disp = Galkin(kwargs_model=kwargs_model, kwargs_aperture=kwargs_aperture, kwargs_psf=kwargs_psf,
+        galkin_analytic = Galkin(kwargs_model=kwargs_model, kwargs_aperture=kwargs_aperture, kwargs_psf=kwargs_psf,
                         kwargs_cosmo=kwargs_cosmo, kwargs_numerics=kwargs_numerics, analytic_kinematics=True)
-        sigma_v2 = los_disp.dispersion(kwargs_mass={'gamma': gamma, 'theta_E': theta_E}, kwargs_light={'r_eff': r_eff},
+        sigma_v_analytic = galkin_analytic.dispersion(kwargs_mass={'gamma': gamma, 'theta_E': theta_E}, kwargs_light={'r_eff': r_eff},
                                        kwargs_anisotropy={'r_ani':r_ani}, sampling_number=1000)
-        print(sigma_v, sigma_v_lin, sigma_v2, 'sigma_v Galkin (log and linear), sigma_v los dispersion')
-        npt.assert_almost_equal(sigma_v2/sigma_v, 1, decimal=2)
+        kwargs_numerics['lum_weight_int_method'] = False
+        galkin_num_3d = Galkin(kwargs_model=kwargs_model, kwargs_aperture=kwargs_aperture, kwargs_psf=kwargs_psf,
+                               kwargs_cosmo=kwargs_cosmo, kwargs_numerics=kwargs_numerics, analytic_kinematics=False)
+        sigma_v_num_3d = galkin_num_3d.dispersion(kwargs_profile, kwargs_light, kwargs_anisotropy, sampling_number=1000)
+
+        print(sigma_v_analytic, sigma_v_num_3d, 'sigma_v Galkin 3d numerics, sigma_v analytic')
+        npt.assert_almost_equal(sigma_v_num_3d/sigma_v_analytic, 1, decimal=2)
+
+        # 2d projected integral calculation
+        kwargs_numerics = {'interpol_grid_num': 1000, 'max_integrate': 1000, 'min_integrate': 0.000001,
+                           'lum_weight_int_method': True, 'log_integration': True}
+        galkin_num_log_proj = Galkin(kwargs_model=kwargs_model, kwargs_aperture=kwargs_aperture, kwargs_psf=kwargs_psf,
+                                kwargs_cosmo=kwargs_cosmo, kwargs_numerics=kwargs_numerics, analytic_kinematics=False)
+        sigma_v_num_log_proj = galkin_num_log_proj.dispersion(kwargs_profile, kwargs_light, kwargs_anisotropy, sampling_number=1000)
+
+        kwargs_numerics = {'interpol_grid_num': 10000, 'max_integrate': 1000, 'min_integrate': 0.0001,
+                           'lum_weight_int_method': True, 'log_integration': False}
+        galkin_num_lin_proj = Galkin(kwargs_model=kwargs_model, kwargs_aperture=kwargs_aperture, kwargs_psf=kwargs_psf,
+                                     kwargs_cosmo=kwargs_cosmo, kwargs_numerics=kwargs_numerics,
+                                     analytic_kinematics=False)
+        sigma_v_num_lin_proj = galkin_num_lin_proj.dispersion(kwargs_profile, kwargs_light, kwargs_anisotropy,
+                                                              sampling_number=1000)
+
+        print(sigma_v_num_log_proj / sigma_v_analytic, sigma_v_num_lin_proj / sigma_v_analytic, 'log proj, lin proj')
+
+        npt.assert_almost_equal(sigma_v_num_log_proj / sigma_v_analytic, 1, decimal=2)
+        npt.assert_almost_equal(sigma_v_num_lin_proj / sigma_v_analytic, 1, decimal=2)
+
+    def test_log_vs_linear_integral(self):
+        """
+        here we test logarithmic vs linear integral in an end-to-end fashion.
+        We do not demand the highest level of precisions here!!!
+        We are using the luminosity-weighted velocity dispersion integration calculation in this test.
+        """
+
+        # light profile
+        light_profile_list = ['HERNQUIST']
+        Rs = .5
+        kwargs_light = [{'Rs':  Rs, 'amp': 1.}]  # effective half light radius (2d projected) in arcsec
+        # 0.551 *
+        # mass profile
+        mass_profile_list = ['SPP']
+        theta_E = 1.2
+        gamma = 2.
+        kwargs_profile = [{'theta_E': theta_E, 'gamma': gamma}]  # Einstein radius (arcsec) and power-law slope
+
+        # anisotropy profile
+        anisotropy_type = 'OM'
+        r_ani = 2.
+        kwargs_anisotropy = {'r_ani': r_ani}  # anisotropy radius [arcsec]
+
+        # aperture as slit
+        aperture_type = 'slit'
+        length = 3.8
+        width = 0.9
+        kwargs_aperture = {'aperture_type': aperture_type, 'length': length, 'width': width, 'center_ra': 0, 'center_dec': 0, 'angle': 0}
+
+        psf_fwhm = 0.7  # Gaussian FWHM psf
+        kwargs_cosmo = {'d_d': 1000, 'd_s': 1500, 'd_ds': 800}
+        kwargs_numerics_log = {'interpol_grid_num': 1000, 'log_integration': True,
+                           'max_integrate': 10, 'min_integrate': 0.001,
+                               'lum_weight_int_method': True}
+        kwargs_numerics_linear = {'interpol_grid_num': 1000, 'log_integration': False,
+                           'max_integrate': 10, 'min_integrate': 0.001,
+                                  'lum_weight_int_method': True}
+        kwargs_psf = {'psf_type': 'GAUSSIAN', 'fwhm': psf_fwhm}
+        kwargs_model = {'mass_profile_list': mass_profile_list,
+                        'light_profile_list': light_profile_list,
+                        'anisotropy_model': anisotropy_type}
+        galkin_linear = Galkin(kwargs_model=kwargs_model, kwargs_aperture=kwargs_aperture, kwargs_psf=kwargs_psf,
+                               kwargs_cosmo=kwargs_cosmo, kwargs_numerics=kwargs_numerics_linear)
+
+        sigma_v_lin = galkin_linear.dispersion(kwargs_profile, kwargs_light, kwargs_anisotropy, sampling_number=1000)
+        galkin_log = Galkin(kwargs_model=kwargs_model, kwargs_aperture=kwargs_aperture, kwargs_psf=kwargs_psf,
+                            kwargs_cosmo=kwargs_cosmo, kwargs_numerics=kwargs_numerics_log)
+        sigma_v_log = galkin_log.dispersion(kwargs_profile, kwargs_light, kwargs_anisotropy, sampling_number=1000)
+        print(sigma_v_lin, sigma_v_log, 'sigma_v linear, sigma_v log')
+        print((sigma_v_lin/sigma_v_log)**2)
+
+        npt.assert_almost_equal(sigma_v_lin/sigma_v_log, 1, decimal=2)
 
     def test_projected_light_integral_hernquist(self):
         """
@@ -215,7 +188,8 @@ class TestGalkin(object):
         :return:
         """
         light_profile_list = ['HERNQUIST']
-        kwargs_light = [{'Rs': 0.10535462602138289, 'center_x': -0.02678473951679429, 'center_y': 0.88691126347462712, 'amp': 3.7114695634960109}]
+        kwargs_light = [{'Rs': 0.10535462602138289, 'center_x': -0.02678473951679429, 'center_y': 0.88691126347462712,
+                         'amp': 3.7114695634960109}]
         lightProfile = LightProfile(light_profile_list)
         R = 0.01
         light2d = lightProfile.light_2d(R=R, kwargs_list=kwargs_light)
@@ -231,7 +205,8 @@ class TestGalkin(object):
         light_profile_list = ['HERNQUIST_ELLIPSE']
         phi, q = 0.74260706384506325, 0.46728323131925864
         e1, e2 = param_util.phi_q2_ellipticity(phi, q)
-        kwargs_light = [{'Rs': 0.10535462602138289, 'e1': e1, 'e2': e2, 'center_x': -0.02678473951679429, 'center_y': 0.88691126347462712, 'amp': 3.7114695634960109}]
+        kwargs_light = [{'Rs': 0.10535462602138289, 'e1': e1, 'e2': e2, 'center_x': -0.02678473951679429,
+                         'center_y': 0.88691126347462712, 'amp': 3.7114695634960109}]
         lightProfile = LightProfile(light_profile_list)
         R = 0.01
         light2d = lightProfile.light_2d(R=R, kwargs_list=kwargs_light)
@@ -309,6 +284,117 @@ class TestGalkin(object):
         sigma_v = galkin.dispersion(kwargs_mass={'theta_E': theta_E, 'gamma': gamma}, kwargs_light={'r_eff': r_eff},
                                     kwargs_anisotropy=kwargs_anisotropy, sampling_number=1000)
         npt.assert_almost_equal(sigma_v, sigma_v_ifu[0], decimal=-1)
+
+    def test_projected_integral_vs_3d_rendering(self):
+
+        lum_weight_int_method = True
+
+        # light profile
+        light_profile_list = ['HERNQUIST']
+        r_eff = 1.5
+        kwargs_light = [{'Rs': 0.551 * r_eff, 'amp': 1.}]  # effective half light radius (2d projected) in arcsec
+        # 0.551 *
+        # mass profile
+        mass_profile_list = ['SPP']
+        theta_E = 1.2
+        gamma = 2.
+        kwargs_profile = [{'theta_E': theta_E, 'gamma': gamma}]  # Einstein radius (arcsec) and power-law slope
+
+        # anisotropy profile
+        anisotropy_type = 'OM'
+        r_ani = 2.
+        kwargs_anisotropy = {'r_ani': r_ani}  # anisotropy radius [arcsec]
+
+        # aperture as slit
+        aperture_type = 'slit'
+        length = 1.
+        width = 0.3
+        kwargs_aperture = {'aperture_type': aperture_type, 'length': length, 'width': width, 'center_ra': 0,
+                           'center_dec': 0, 'angle': 0}
+
+        psf_fwhm = 1.  # Gaussian FWHM psf
+        kwargs_cosmo = {'d_d': 1000, 'd_s': 1500, 'd_ds': 800}
+        kwargs_numerics_3d = {'interpol_grid_num': 2000, 'log_integration': True,
+                           'max_integrate': 1000, 'min_integrate': 0.00001, 'lum_weight_int_method': False
+                           }
+        kwargs_model = {'mass_profile_list': mass_profile_list,
+                        'light_profile_list': light_profile_list,
+                        'anisotropy_model': anisotropy_type}
+        kwargs_psf = {'psf_type': 'GAUSSIAN', 'fwhm': psf_fwhm}
+        galkin = Galkin(kwargs_model=kwargs_model, kwargs_aperture=kwargs_aperture, kwargs_psf=kwargs_psf,
+                        kwargs_cosmo=kwargs_cosmo, kwargs_numerics=kwargs_numerics_3d)
+        sigma_v = galkin.dispersion(kwargs_profile, kwargs_light, kwargs_anisotropy, sampling_number=1000)
+
+        kwargs_numerics_2d = {'interpol_grid_num': 2000, 'log_integration': True,
+                              'max_integrate': 1000, 'min_integrate': 0.00001, 'lum_weight_int_method': True
+                              }
+
+        galkin = Galkin(kwargs_model=kwargs_model, kwargs_aperture=kwargs_aperture, kwargs_psf=kwargs_psf,
+                        kwargs_cosmo=kwargs_cosmo, kwargs_numerics=kwargs_numerics_2d, analytic_kinematics=False)
+        sigma_v_int_method = galkin.dispersion(kwargs_profile, kwargs_light, kwargs_anisotropy, sampling_number=1000)
+        npt.assert_almost_equal(sigma_v_int_method / sigma_v, 1, decimal=2)
+
+    def test_2d_vs_3d_power_law(self):
+        # set up power-law light profile
+        light_model = ['POWER_LAW']
+        kwargs_light = [{'gamma': 2, 'amp': 1, 'e1': 0, 'e2': 0}]
+
+        lens_model = ['SIS']
+        kwargs_mass = [{'theta_E': 1}]
+
+        anisotropy_type = 'isotropic'
+        kwargs_anisotropy = {}
+        kwargs_model = {'mass_profile_list': lens_model,
+                        'light_profile_list': light_model,
+                        'anisotropy_model': anisotropy_type}
+        kwargs_numerics = {'interpol_grid_num': 2000, 'log_integration': True,
+                           'max_integrate': 50, 'min_integrate': 0.0001}
+
+        kwargs_numerics_3d = copy.deepcopy(kwargs_numerics)
+        kwargs_numerics_3d['lum_weight_int_method'] = False
+
+        kwargs_numerics_2d = copy.deepcopy(kwargs_numerics)
+        kwargs_numerics_2d['lum_weight_int_method'] = True
+
+        kwargs_cosmo = {'d_d': 1000, 'd_s': 1500, 'd_ds': 800}
+
+        # compute analytic velocity dispersion of SIS profile
+
+        v_sigma_c2 = kwargs_mass[0]['theta_E'] * const.arcsec / (4 * np.pi) * kwargs_cosmo['d_s'] / kwargs_cosmo['d_ds']
+        v_sigma_true = np.sqrt(v_sigma_c2) * const.c / 1000
+
+        # aperture as slit
+        aperture_type = 'slit'
+        length = 1.
+        width = 0.3
+        kwargs_aperture = {'aperture_type': aperture_type, 'length': length, 'width': width, 'center_ra': 0,
+                           'center_dec': 0, 'angle': 0}
+        kwargs_psf = {'psf_type': 'GAUSSIAN', 'fwhm': 0.5}
+
+        galkin3d = Galkin(kwargs_model=kwargs_model, kwargs_aperture=kwargs_aperture, kwargs_psf=kwargs_psf,
+                        kwargs_cosmo=kwargs_cosmo, kwargs_numerics=kwargs_numerics_3d)
+
+        galkin2d = Galkin(kwargs_model=kwargs_model, kwargs_aperture=kwargs_aperture, kwargs_psf=kwargs_psf,
+                          kwargs_cosmo=kwargs_cosmo, kwargs_numerics=kwargs_numerics_2d)
+
+        sigma_draw_list = []
+        for i in range(100):
+            sigma_v_draw = galkin3d._draw_one_sigma2(kwargs_mass, kwargs_light, kwargs_anisotropy)
+            sigma_draw_list.append(sigma_v_draw)
+            # print(np.sqrt(sigma_v_draw)/ 1000)
+
+        #import matplotlib.pyplot as plt
+        #plt.plot(np.sqrt(sigma_draw_list) / 1000 / v_sigma_true)
+        #plt.show()
+
+        print(np.sqrt(np.mean(sigma_draw_list)) / 1000, 'mean draw')
+        print('truth = ', v_sigma_true)
+        #assert 1 == 0
+
+        sigma_v_2d = galkin2d.dispersion(kwargs_mass, kwargs_light, kwargs_anisotropy, sampling_number=1000)
+        sigma_v_3d = galkin3d.dispersion(kwargs_mass, kwargs_light, kwargs_anisotropy, sampling_number=1000)
+        npt.assert_almost_equal(sigma_v_2d / v_sigma_true, 1, decimal=2)
+        npt.assert_almost_equal(sigma_v_3d / v_sigma_true, 1, decimal=2)
 
 
 class TestRaise(unittest.TestCase):